requirements.txt
hydra-core
pyautogui